Extract xarray geoextensions
These are used throughout the codebase, and should be visibly separate.
