+++ datacube/__init__.py
@@ -6,7 +6,7 @@ Provides access to multi-dimensional data, with a focus on Earth observations da
 
 To use this module, see the `Developer Guide <http://datacube-core.readthedocs.io/en/stable/dev/developer.html>`_.
 
-The main class to access the datacube is :py:class:`datacube.Datacube`.
+The main class to access the datacube is :class:`datacube.Datacube`.
 
 To initialise this class, you will need a config pointing to a database, such as a file with the following::
 
@@ -21,5 +21,8 @@ from .version import __version__
 from .api import Datacube
 from .config import set_options
 import warnings
+from .utils import xarray_geoextensions
 
+
+# Ensure deprecation warnings from datacube modules are shown
 warnings.filterwarnings('always', category=DeprecationWarning, module=r'^datacube\.')
